Rework initialization with unset resources. Start packaging for Windows installable app.
